Revert "Allow httprewrite.BufferRewrite to return an error (#318)"
This reverts commit e9a47585f73a4cc4137f454154d3c77cc9870633.
